Fix memory leak
When the editor is initialized, we attach an object to the editor instance and to the DOM. That object contains a reference to the editor, and the editor contains a reference to the DOM node. This produces a reference loop that isn't cleaned up on destroy. This fixes that by deleting the object reference on the DOM.